Pass around JSON in drag-and-drop
We can't pass raw objects, but JSON-encoded text manages to slip
through on Chrome+Firefox. HOWEVER, all text is lowercased in the
process, so we're limited to underscore-separation for many things.

This is all because dragover events can't get at the attached data on
the drag event, which is desperately needed when we're accepting/
rejecting drops based on criteria more complicated than a goddamned
mimetype.