Reorganize
Trying to declutter the client JS stuff by merging a few things,
renaming more things to be pane- as appropriate, and making all
the entry points be main.js/main.jsx. As we're not using ES6 classes
anymore, there's no point to having displayName set, so those are
stabbed as well.