More precise column resizing
The visual indicator didn't always exactly line up with the intended width
because it displayed at the mouse position and the width was just changed
by the diff of the mouse position. Now the width is determined directly by
the mouse position rather than just diffed by the mouse diff, so it should
be exact.